Add registry & blob cache checker to CRI pull
Introduces a new image handler during CRI image pull. This handler keeps track of which
blobs are pulled from which registries and skips the image pull *only* if the same blob is
pulled from the same registry before. If the blob is same but is being pulled from a
different registry then we still go ahead and pull it to ensure that the registry actually
has the correct blob.

Signed-off-by: Amit Barve <ambarve@microsoft.com>