Add image pull records handler wrapper
Earlier commit added image pull records handler as a normal ImageHandler that would run
along with other ImageHandlers. However, this doesn't allow us to control the entire blob
download operation. i.e we want to be able to add a new record only if the blob download
is successful. For this we need an ImageHandlerWrapper that wraps the ImageHandlers and
then does some pre/post processing based on the results of ImageHandlers.
This commit modifies the pull records ImageHandler to make it a ImageHandlerWrapper.

Signed-off-by: Amit Barve <ambarve@microsoft.com>